+++ src/translations/en-GB/gtag.php
@@ -5,5 +5,6 @@ return [
 	'Example : UA-123456789-1. Entering no value here will disable gtags' => 'Example : UA-123456789-1. Entering no value here will disable gtags',
 	'Only in production mode' => 'Only in production mode',
 	'Settings for {site}' => 'Settings for {site}',
-	'Add a site before changing these settings' => 'Add a site before changing these settings'
+	'Add a site before changing these settings' => 'Add a site before changing these settings',
+	'Flags added to the _ga cookie' => 'Flags added to the _ga cookie'
 ];