+++ src/translations/fr/gtag.php
@@ -5,5 +5,6 @@ return [
 	'Example : UA-123456789-1. Entering no value here will disable gtags' => 'Example : UA-123456789-1. Pas de valeur ici desactivera gtags',
 	'Only in production mode' => 'Seulement en mode production',
 	'Settings for {site}' => 'Paramètres pour {site}',
-	'Add a site before changing these settings' => 'Ajoutez un site avant de modifier ces paramètres'
+	'Add a site before changing these settings' => 'Ajoutez un site avant de modifier ces paramètres',
+	'Flags added to the _ga cookie' => 'Indicateurs ajoutés au cookie _ga'
 ];